requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
Pillow>=9.5.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.2.0
matplotlib>=3.7.0
seaborn>=0.12.0
pyyaml>=6.0
tqdm>=4.65.0
flask>=2.3.0
flask-cors>=4.0.0
opencv-python>=4.7.0
grad-cam>=1.4.8
tensorboard>=2.13.0
datasets>=2.12.0
streamlit>=1.28.0
easyocr>=1.7.0
plotly>=5.15.0
